2020/12/23/09.38

+++ src/main/java/service/shopCarservice.java
@@ -28,39 +28,42 @@ public class shopCarservice{
 	public List<Product> getData(String doWhich, Integer show, Integer  buyHowmuch, Integer memberId, Integer productId) {
 		List<Product> products = new ArrayList<Product>();
 		List<ShopCar> lShopCar = new ArrayList<ShopCar>();
-		switch (show) {
-			case -1:
-				return productDao.SearchAllGame();
+		if(show == -1) {
+			return productDao.SearchAllGame();
+		}
+		if(memberId != null) {
+			switch (show) {
 			case 0:
 				switch (doWhich) {
-					case "":
-						lShopCar = shopCarDao.selectAll(memberId);
-						for(ShopCar s : lShopCar) {
-							products.add(s.getpId());
-						}
-						return products;
-					case "insert":
-						Product product = productDao.SearchGame(productId);
-						shopCarDao.insert(new ShopCar(product, memberDao.getMember(memberId), buyHowmuch, "N"));
-						products.add(product);
-						return products;
-					case "update":
-						shopCarDao.update(memberId, productId, buyHowmuch);
-						return products;
-					case "delete":
-						shopCarDao.delete(memberId, productId);
-						lShopCar = shopCarDao.selectAll(memberId);
-						for(ShopCar s : lShopCar) {
-							products.add(s.getpId());
-						}
-						return products;
+				case "":
+					lShopCar = shopCarDao.selectAll(memberId);
+					for(ShopCar s : lShopCar) {
+						products.add(s.getpId());
+					}
+					return products;
+				case "insert":
+					Product product = productDao.SearchGame(productId);
+					shopCarDao.insert(new ShopCar(product, memberDao.getMember(memberId), buyHowmuch, "N"));
+					products.add(product);
+					return products;
+				case "update":
+					shopCarDao.update(memberId, productId, buyHowmuch);
+					return products;
+				case "delete":
+					shopCarDao.delete(memberId, productId);
+					lShopCar = shopCarDao.selectAll(memberId);
+					for(ShopCar s : lShopCar) {
+						products.add(s.getpId());
+					}
+					return products;
 				}
 			case 1:
 				break;
 			case 2:
 				break;
+			}
 		}
-		return null;
+		return products;
 	}
 	@Transactional
 	public Map<Integer, Integer> getquantity(Integer memberId){
